Adapt the output folders to handle multi config generators
The Visual studio generator is a multi config generator.
This means, that form one project, e.g debug and release can be build.
This caused issues, as the default behaviour is to append the configuration type to the output directory.
Thus, dlls and required files were not copied to the right locations.

By adding the generator expression '$<Config>' to the CMAKE_RUNTIME_OUTPUT_DIRECTORY path,
the default appending behaviour is supressed and every configuration uses its own output folder.
The dlls and required files can then be copied in the correct folder for each configuration.